requirements.txt
python-dotenv
requests
ipykernel
numpy
langchain_community
pypdf
langchain 
sentence-transformers 
pinecone-client
pandas 
matplotlib 
seaborn
langchain-pinecone
pyMuPDF
python-box
streamlit
langgraph
langchain-groq
rank-bm25